Reintroduce logseq-pages variable
This variable was removed in commit
7355dfb646b188c0612a1a24c18bc378e41dffc9

This is essentially just an alias for the variable
`logseq-id-title-mod-path` which bears the
confusing description "paths where id and
title additions are allowed". I think this
is a bit odd considering it is just a single
path. I can perhaps add more in the future.

The `mod-path` variable was added in commit
dbf433c523f045d4e2afc990f97608c958602c86