Fix logical error regression
See https://github.com/FrederickGeek8/org-roam-logseq.el/pull/1
for more information.

There is a small bug with the boolean logic that
was introduced in a previous commit as a "fix".
I think it can be a bit confusing because we
trigger a null-action if the if statement is true,
indicating that we should skip processing the file.

Because of this bug, enabling journal modification
would not take effect, since the journal is not
"in the search path" that is checked in the
second part of the "or" statement.

We want:
- if (skip-journal and not-path) skip

Instead we have
- if (skip-journal or not-path) skip

If we wanted to include a journal (skip-journal=False)
then it would be excluded if the journals are not
in the "search path".

This bug was introduced in commit
0f7c85e43fcefade7405d1792c12633586fb4465